Added human_detector class and it's header file
This derives the previous CameraModel class and communicates with it to localize the pixels in 3D coordinates w.r.t camera frame. For now, the calculation is done by manual bounding boxes around subject of interest and choosing interested pixel.

This commit also makes changes to CMakeList.txt file to run everything together